Seperating out modules, use of fragmentation, adding editAuthor functionality
