+++ src/components/authors/AuthorsPage.tsx
@@ -1,0 +1,9 @@
+import AuthorList from "@/components/authors/AuthorList";
+
+export default function AuthorsPage() {
+    return (
+        <div className="container mx-auto py-8">
+            <AuthorList />
+        </div>
+    );
+}
