+++ src/components/books/BooksPage.tsx
@@ -1,0 +1,9 @@
+import BookList from "@/components/books/BookList";
+
+export default function BooksPage() {
+    return (
+        <div className="container mx-auto py-8">
+            <BookList />
+        </div>
+    );
+}
