+++ src/types/author.d.ts
@@ -1,0 +1,9 @@
+import { Book } from "./book";
+
+export interface Author {
+    id: string;
+    name: string;
+    biography: string;
+    bornDate: string;
+    books?: Book[];
+}
