+++ src/types/book.d.ts
@@ -1,0 +1,9 @@
+import { Author } from "./author";
+
+export interface Book {
+    id: string;
+    title: string;
+    description: string;
+    published_date: string;
+    author?: Author;
+}
